+++ internal/fasttime/fasttime_x64.go
@@ -2,8 +2,22 @@
 
 package fasttime
 
-import "strconv"
+import (
+	"fmt"
+	"strconv"
+	"strings"
+)
 
-// As int on 64-bit systems is 64 bit, it is possible to use faster Atoi.
-
-var atoi = strconv.Atoi
+// TS2int converts a slack timestamp to an int64 by stripping the dot and
+// converting the string to an int64.  It is useful for fast comparison.
+func TS2int(ts string) (int64, error) {
+	if ts == "" {
+		return 0, nil
+	}
+	i := strings.IndexByte(ts, '.')
+	if i == -1 {
+		return 0, fmt.Errorf("%w: %q", ErrNotATimestamp, ts)
+	}
+	val, err := strconv.Atoi(ts[:i] + ts[i+1:])
+	return int64(val), err
+}
